+++ gobblin-yarn/src/main/java/org/apache/gobblin/yarn/GobblinApplicationMaster.java
@@ -127,7 +127,7 @@ public class GobblinApplicationMaster extends GobblinClusterManager {
    * Build the {@link YarnContainerSecurityManager} for the Application Master.
    */
   private YarnContainerSecurityManager buildYarnContainerSecurityManager(Config config, FileSystem fs) {
-    return new YarnContainerSecurityManagerForApplicationMaster(config, fs, this.eventBus, this.yarnService);
+    return new YarnMasterSecurityManager(config, fs, this.eventBus, this.yarnService);
   }
 
   @Override
